+++ packages/databricks-vscode/CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## 1.4.0 (2024-08-28)
 
+-   Added a notification about upcoming V2 release:
+    > **Please note the V2 version of the extension will be made the default in a few days. With this update it is easier to set up your project, integrate with Databricks Asset Bundles and run your code remotely with fewer steps. [Learn more](https://docs.databricks.com/en/dev-tools/vscode-ext/install.html).**
+
 # Release: v1.3.1
 
 ## packages/databricks-vscode